pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "accessible-pdf-toolkit"
version = "1.0.0"
description = "Privacy-first desktop application for making PDFs WCAG 2.1/2.2 compliant using local AI models"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Accessible PDF Toolkit Team"}
]
keywords = [
    "pdf",
    "accessibility",
    "wcag",
    "pdf-ua",
    "education",
    "ferpa",
    "hipaa",
    "local-ai",
    "privacy"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: X11 Applications :: Qt",
    "Intended Audience :: Education",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Text Processing :: Markup",
    "Topic :: Multimedia :: Graphics :: Graphics Conversion",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "PyQt6>=6.4.0",
    "PyMuPDF>=1.23.0",
    "pikepdf>=8.0.0",
    "Pillow>=10.0.0",
    "pytesseract>=0.3.10",
    "cryptography>=41.0.0",
    "bcrypt>=4.0.0",
    "SQLAlchemy>=2.0.0",
    "requests>=2.31.0",
    "httpx>=0.25.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "pyyaml>=6.0",
    "pyqtgraph>=0.13.0",
    "gpt4all>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-qt>=4.2.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
accessible-pdf-toolkit = "accessible_pdf_toolkit.main:main"

[project.urls]
Homepage = "https://github.com/accessible-pdf-toolkit/accessible-pdf-toolkit"
Documentation = "https://github.com/accessible-pdf-toolkit/accessible-pdf-toolkit/tree/main/docs"
Repository = "https://github.com/accessible-pdf-toolkit/accessible-pdf-toolkit"
Issues = "https://github.com/accessible-pdf-toolkit/accessible-pdf-toolkit/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
accessible_pdf_toolkit = ["assets/**/*"]

[tool.black]
line-length = 88
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=accessible_pdf_toolkit --cov-report=term-missing"
